requirements.txt
# core
pandas>=2.1
numpy>=1.26
pyarrow>=15

# geospatial
geopandas>=0.14
pyproj>=3.6
shapely>=2.0
fiona>=1.9
rasterio>=1.3
rioxarray>=0.15
xarray>=2024.1
rasterstats>=0.19
pyogrio>=0.7

# earth engine (for automated VIIRS download)
earthengine-api>=0.1.390

# cli + utils
typer>=0.12
rich>=13.7
tqdm>=4.66
pyyaml>=6.0
requests>=2.31
